Preprocess input

+++ src/main/java/edu/wsu/backendapi/controller/TemplateController.java
@@ -13,9 +13,7 @@ import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.sql.SQLException;
 
-import org.json.JSONObject;
 import org.apache.http.HttpResponse;
-
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -26,70 +24,68 @@ public class TemplateController {
     @GET
     @Path("/siteflow/get/all")
     @Produces(MediaType.APPLICATION_JSON)
-    public Object templateGetAll() {
+    public Response templateGetAll() {
         SiteflowService tempServ = new SiteflowService();
         try {
             HttpResponse output = tempServ.getAllTemplates();
-            return printInfo(output,true);
+            return Response.ok(printInfo(output,true)).build();
         } catch (InvalidKeyException | NoSuchAlgorithmException | IOException e) {
             e.printStackTrace();
+            return Response.status(400,"Error").build();
         }
-
-
-        JSONObject obj = new JSONObject();
-        obj.put("status", 400);
-        return obj.toString(4);
     }
 
     @GET
     @Path("/rds/get/all")
     @Produces(MediaType.APPLICATION_JSON)
-    public Object templateGetAllRds() {
+    public Response templateGetAllRds() {
         TemplateDao tempGetRds = new TemplateDao();
         try {
-            return tempGetRds.getTemplateAllRds();
-        } catch ( SQLException e) {
+            return Response.ok(tempGetRds.getTemplateAllRds().toString()).build();
+        } catch (SQLException e) {
             e.printStackTrace();
+            return Response.status(400,"SQL Error").build();
         }
-
-        JSONObject obj = new JSONObject();
-        obj.put("status", 400);
-        return obj.toString(4);
     }
 
     @GET
     @Path("/rds/get")
     @Produces(MediaType.APPLICATION_JSON)
-    public Object templateGetById(@QueryParam("id") int id) {
+    public Response templateGetById(@QueryParam("id") int id) {
         TemplateDao tempGetRds = new TemplateDao();
+
         try {
-            return tempGetRds.getTemplateByIdRds(id);
+            if (tempGetRds.checkExist(id)) {
+                try {
+                    return Response.ok(tempGetRds.getTemplateByIdRds(id).toString()).build();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                    return Response.status(400, "SQL Error").build();
+                }
+            } else {
+                return Response.status(204, "Content Does Not Exist, No Content").build();
+            }
         } catch (SQLException e) {
             e.printStackTrace();
+            return Response.status(400, "SQL Error").build();
         }
-
-        JSONObject obj = new JSONObject();
-        obj.put("id",id);
-        obj.put("Error Message","Not Found");
-        return obj.toString(4);
     }
 
     @DELETE
     @Path("/rds/delete")
     @Produces(MediaType.TEXT_PLAIN)
     public Response templateDeleteById(@QueryParam("id") int id) {
-        int retVal = 0;
         TemplateDao tempDeleteRds = new TemplateDao();
         try {
-            retVal = tempDeleteRds.deleteTemplateByIdRds(id);
+            if (tempDeleteRds.checkExist(id)) {
+                tempDeleteRds.deleteTemplateByIdRds(id);
+                return Response.status(200, "Template Deleted Successfully").build();
+            } else {
+                return Response.status(204, "Template Not Deleted, No Content").build();
+            }
         } catch (SQLException e) {
             e.printStackTrace();
-        }
-
-        if (retVal == 1) {
-            return Response.status(Response.Status.OK).build();
-        } else {
-            return Response.status(Response.Status.NO_CONTENT).build();
+            return Response.status(400,"SQL Error").build();
         }
     }
 
@@ -97,48 +93,72 @@ public class TemplateController {
     @Path("/rds/post")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.TEXT_PLAIN)
-    public String templatePostDb(String jsonIn) {
+    public Response templatePostDb(String jsonIn) {
         ObjectMapper objectMapper = new ObjectMapper();
         //objectMapper.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
-        Template templateRdsUpdate = new Template();
+        Template templateRdsUpdate;
         try {
             templateRdsUpdate = objectMapper.readValue(jsonIn, Template.class);
 
         } catch (JsonProcessingException e) {
             e.printStackTrace();
+            return Response.status(400,"JSON Mapping Error").build();
         }
 
         PreProcess preprocessInput = new PreProcess();
-        int isGood = preprocessInput.PreProcessTemplate(templateRdsUpdate);
-        System.out.println("isGood: " + isGood);
-
-        TemplateDao addTempDao = new TemplateDao();
-        addTempDao.addTemplate(templateRdsUpdate);
-        return "received";
+        if (preprocessInput.PreProcessTemplate(templateRdsUpdate)) {
+            TemplateDao addTempDao = new TemplateDao();
+            try {
+                return Response.ok(addTempDao.addTemplate(templateRdsUpdate).toString()).build();
+            }catch (SQLException e) {
+                e.printStackTrace();
+                return Response.status(400,"SQL Error").build();
+            } catch (JsonProcessingException e) {
+                e.printStackTrace();
+                return Response.status(400,"JSON Processing Error").build();
+            }
+        } else {
+            return Response.status(400,"Preprocessing Error").build();
+        }
     }
 
     @PUT
     @Path("/rds/update")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.TEXT_PLAIN)
-    public String templateUpdateDb(String jsonIn, @QueryParam("id") int idIn) {
+    public Response templateUpdateDb(String jsonIn, @QueryParam("id") int idIn) {
+        int retVal = 0;
         ObjectMapper objectMapper = new ObjectMapper();
         //objectMapper.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
-        Template templateRdsUpdate = new Template();
+        Template templateRdsUpdate;
         try {
             templateRdsUpdate = objectMapper.readValue(jsonIn, Template.class);
 
         } catch (JsonProcessingException e) {
             e.printStackTrace();
+            return Response.status(400,"JSON Mapping Error").build();
         }
 
         PreProcess preprocessInput = new PreProcess();
-        int isGood = preprocessInput.PreProcessTemplate(templateRdsUpdate);
-        System.out.println("isGood: " + isGood);
+        if (preprocessInput.PreProcessTemplate(templateRdsUpdate)) {
+            TemplateDao updateTempDao = new TemplateDao();
+            try {
+                if (updateTempDao.checkExist(idIn)) {
+                    updateTempDao.updateTemplate(templateRdsUpdate, idIn);
+                    return Response.status(200, "Template Updated Successfully").build();
+                } else {
+                    return Response.status(204, "Template Update Unsuccessful, No Content Item").build();
+                }
+            } catch (SQLException e) {
+                e.printStackTrace();
+                return Response.status(400,"SQL Error").build();
+            } catch (JsonProcessingException e) {
+                e.printStackTrace();
+                return Response.status(400,"JSON Processing Error").build();
+            }
 
-        TemplateDao addTempDao = new TemplateDao();
-        addTempDao.updateTemplate(templateRdsUpdate, idIn);
-
-        return "received";
+        } else {
+            return Response.status(400,"Preprocessing Error").build();
+        }
     }
 }
